+++ src/components/mailFilter/MailFilterActionFileinto.vue
@@ -11,7 +11,7 @@
 <script>
 
 import NcSelect from '@nextcloud/vue/dist/Components/NcSelect.js'
-import { mailboxHasRights } from '../util/acl.js'
+import { mailboxHasRights } from '../../util/acl.js'
 export default {
 	name: 'MailFilterActionFileinto',
 	components: {
